CC #32 - Renaming AutocompleteDropdown to RemoteDropdown
