+++ src/index.js
@@ -2,11 +2,10 @@
 
 // Common
 export { default as Keyboard } from './common/Keyboard';
-export { default as TagsList} from './common/TagsList';
+export { default as TagsList } from './common/TagsList';
 
 // Semantic UI
 export { default as AssociatedDropdown } from './semantic-ui/AssociatedDropdown';
-export { default as AutocompleteDropdown } from './semantic-ui/AutocompleteDropdown';
 export { default as BooleanIcon } from './semantic-ui/BooleanIcon';
 export { default as ColorButton } from './semantic-ui/ColorButton';
 export { default as ColorPickerModal } from './semantic-ui/ColorPickerModal';
@@ -24,6 +23,7 @@ export { default as LinkLabel } from './semantic-ui/LinkLabel';
 export { default as ListTable } from './semantic-ui/ListTable';
 export { default as LoginModal } from './semantic-ui/LoginModal';
 export { default as PhotoViewer } from './semantic-ui/PhotoViewer';
+export { default as RemoteDropdown } from './semantic-ui/RemoteDropdown';
 export { default as Selectize } from './semantic-ui/Selectize';
 export { default as TabbedModal } from './semantic-ui/TabbedModal';
